requirements.txt
tweepy==4.14.0
feedparser==6.0.11
requests==2.32.3
python-dotenv==1.0.1
